minor tune
